+++ app/models/carto/group.rb
@@ -2,9 +2,21 @@
 
 require 'active_record'
 require_dependency 'cartodb/errors'
+require_relative 'paged_model'
 
 module Carto
+
+  # Groups are created by the editor because of extension requests, so
+  # standard Rails operations (creation, destruction, etc) doesn't trigger
+  # extension management functions. In order to keep extension and database
+  # in sync, there're several methods that do trigger it:
+  # - create_group_with_extension
+  # - rename_group_with_extension
+  # - destroy_group_with_extension
+  # - add_member_with_extension
+  # - remove_member_with_extension
   class Group < ActiveRecord::Base
+    include PagedModel
 
     belongs_to :organization, class_name: Carto::Organization
     has_many :users_group, dependent: :destroy, class_name: Carto::UsersGroup
@@ -14,6 +26,7 @@ module Carto
 
     validates :name, :database_role, :organization_id, :presence => true
 
+    # Constructor for groups already existing in the database
     def self.new_instance(database_name, name, database_role, display_name = name)
       organization = Organization.find_by_database_name(database_name)
 
@@ -22,6 +35,53 @@ module Carto
       raise CartoDB::ModelAlreadyExistsError if Group.find_by_organization_id_and_name_and_database_role(organization.id, name, database_role)
 
       new(name: name, database_role: database_role, display_name: display_name, organization: organization)
+    end
+
+    # Creation of brand-new group with the extension
+    def self.create_group_with_extension(organization, display_name)
+      name = valid_group_name(display_name)
+      organization.owner.in_database do |conn|
+        create_group_extension_query(conn, name)
+      end
+      # Extension triggers a request to the editor databases endpoint which actually creates the group
+      group = Carto::Group.find_by_organization_id_and_name(organization.id, name)
+      raise "Group was not created by the extension. Is it installed and configured?" if group.nil?
+
+      group.display_name = display_name
+      group.save
+      group
+    end
+
+    def rename_group_with_extension(new_display_name)
+      new_name = Carto::Group.valid_group_name(new_display_name)
+      organization.owner.in_database do |conn|
+        Carto::Group.rename_group_extension_query(conn, name, new_name)
+      end
+      self.reload
+      self.display_name = new_display_name
+      self.save
+    end
+
+    # INFO: public because it's called by Organization.
+    def destroy_group_with_extension
+      # INFO: currently only a superuser can destroy a group. See CartoDB/cartodb-postgresql#114
+      organization.owner.in_database(as: :superuser) do |conn|
+        Carto::Group.destroy_group_extension_query(conn, name)
+      end
+    end
+
+    def add_member_with_extension(user)
+      organization.owner.in_database do |conn|
+        Carto::Group.add_member_group_extension_query(conn, name, user.username)
+      end
+      self.reload
+    end
+
+    def remove_member_with_extension(user)
+      organization.owner.in_database do |conn|
+        Carto::Group.remove_member_group_extension_query(conn, name, user.username)
+      end
+      self.reload
     end
 
     def rename(new_name, new_database_role)
@@ -51,6 +111,35 @@ module Carto
 
     def database_name
       organization.database_name
+    end
+
+    private
+
+    # TODO: PG Format("%I", strvar); ?
+    def self.valid_group_name(display_name)
+      name = display_name.squish
+      name = "g_#{name}" unless name[/^[a-zA-Z_]{1}/]
+      name.gsub(/[^a-zA-Z0-9_]/,'_').gsub(/_{2,}/, '_')
+    end
+
+    def self.create_group_extension_query(conn, name)
+      conn.execute(%Q{ select cartodb.CDB_Group_CreateGroup('#{name}') })
+    end
+
+    def self.rename_group_extension_query(conn, name, new_name)
+      conn.execute(%Q{ select cartodb.CDB_Group_RenameGroup('#{name}', '#{new_name}') })
+    end
+
+    def self.destroy_group_extension_query(conn, name)
+      conn.execute(%Q{ select cartodb.CDB_Group_DropGroup('#{name}') })
+    end
+
+    def self.add_member_group_extension_query(conn, name, username)
+      conn.execute(%Q{ select cartodb.CDB_Group_AddMember('#{name}', '#{username}') })
+    end
+
+    def self.remove_member_group_extension_query(conn, name, username)
+      conn.execute(%Q{ select cartodb.CDB_Group_RemoveMember('#{name}', '#{username}') })
     end
 
   end