+++ config/environments/development.rb
@@ -44,7 +44,8 @@ CartoDB::Application.configure do
   # config.action_mailer.delivery_method = :smtp
 
   # Enable threaded mode
-  # config.threadsafe!
+  # In order to run supporting concurrent requests, uncomment next line and run with `bundle exec thin start --threaded -p 3000`
+  config.threadsafe! unless $rails_rake_task
 
   # Enable locale fallbacks for I18n (makes lookups for any locale fall back to
   # the I18n.default_locale when a translation can not be found)